requirements.txt
Flask==0.10.1
future==0.16.0
paho-mqtt==1.1